+++ fastly/block_fastly_service_v1_logging_digitalocean_test.go
@@ -20,7 +20,7 @@ func TestResourceFastlyFlattenDigitalOcean(t *testing.T) {
 		{
 			remote: []*gofastly.DigitalOcean{
 				{
-					Version:           1,
+					ServiceVersion:    1,
 					Name:              "digitalocean-endpoint",
 					BucketName:        "bucket",
 					AccessKey:         "access",
@@ -74,7 +74,7 @@ func TestAccFastlyServiceV1_logging_digitalocean_basic(t *testing.T) {
 	domain := fmt.Sprintf("fastly-test.%s.com", name)
 
 	log1 := gofastly.DigitalOcean{
-		Version:           1,
+		ServiceVersion:    1,
 		Name:              "digitalocean-endpoint",
 		BucketName:        "bucket",
 		AccessKey:         "access",
@@ -93,7 +93,7 @@ func TestAccFastlyServiceV1_logging_digitalocean_basic(t *testing.T) {
 	}
 
 	log1_after_update := gofastly.DigitalOcean{
-		Version:           1,
+		ServiceVersion:    1,
 		Name:              "digitalocean-endpoint",
 		BucketName:        "bucketupdate",
 		AccessKey:         "accessupdate",
@@ -112,7 +112,7 @@ func TestAccFastlyServiceV1_logging_digitalocean_basic(t *testing.T) {
 	}
 
 	log2 := gofastly.DigitalOcean{
-		Version:           1,
+		ServiceVersion:    1,
 		Name:              "another-digitalocean-endpoint",
 		BucketName:        "bucket2",
 		AccessKey:         "access2",
@@ -168,7 +168,7 @@ func TestAccFastlyServiceV1_logging_digitalocean_basic_compute(t *testing.T) {
 	domain := fmt.Sprintf("fastly-test.%s.com", name)
 
 	log1 := gofastly.DigitalOcean{
-		Version:         1,
+		ServiceVersion:  1,
 		Name:            "digitalocean-endpoint",
 		BucketName:      "bucket",
 		AccessKey:       "access",
@@ -207,8 +207,8 @@ func testAccCheckFastlyServiceV1DigitalOceanAttributes(service *gofastly.Service
 
 		conn := testAccProvider.Meta().(*FastlyClient).conn
 		digitaloceanList, err := conn.ListDigitalOceans(&gofastly.ListDigitalOceansInput{
-			Service: service.ID,
-			Version: service.ActiveVersion.Number,
+			ServiceID:      service.ID,
+			ServiceVersion: service.ActiveVersion.Number,
 		})
 
 		if err != nil {
@@ -226,7 +226,7 @@ func testAccCheckFastlyServiceV1DigitalOceanAttributes(service *gofastly.Service
 				if e.Name == el.Name {
 					// we don't know these things ahead of time, so populate them now
 					e.ServiceID = service.ID
-					e.Version = service.ActiveVersion.Number
+					e.ServiceVersion = service.ActiveVersion.Number
 					// We don't track these, so clear them out because we also wont know
 					// these ahead of time
 					el.CreatedAt = nil