+++ fastly/block_fastly_service_v1_requestsetting_test.go
@@ -95,8 +95,8 @@ func testAccCheckFastlyServiceV1RequestSettingsAttributes(service *gofastly.Serv
 
 		conn := testAccProvider.Meta().(*FastlyClient).conn
 		rqList, err := conn.ListRequestSettings(&gofastly.ListRequestSettingsInput{
-			Service: service.ID,
-			Version: service.ActiveVersion.Number,
+			ServiceID:      service.ID,
+			ServiceVersion: service.ActiveVersion.Number,
 		})
 
 		if err != nil {
@@ -113,7 +113,7 @@ func testAccCheckFastlyServiceV1RequestSettingsAttributes(service *gofastly.Serv
 				if r.Name == lr.Name {
 					// we don't know these things ahead of time, so populate them now
 					r.ServiceID = service.ID
-					r.Version = service.ActiveVersion.Number
+					r.ServiceVersion = service.ActiveVersion.Number
 					// We don't track these, so clear them out because we also wont know
 					// these ahead of time
 					lr.CreatedAt = nil