+++ fastly/resource_fastly_service_dictionary_items_v1_test.go
@@ -63,7 +63,44 @@ func TestAccFastlyServiceDictionaryItemV1_create(t *testing.T) {
 		CheckDestroy: testAccCheckServiceV1Destroy,
 		Steps: []resource.TestStep{
 			{
-				Config: testAccServiceDictionaryItemsV1Config_one_dictionary_with_items(name, dictName, expectedRemoteItems),
+				Config: testAccServiceDictionaryItemsV1Config_one_dictionary_with_items(name, dictName, expectedRemoteItems, true),
+				Check: resource.ComposeTestCheckFunc(
+					testAccCheckServiceV1Exists("fastly_service_v1.foo", &service),
+					testAccCheckFastlyServiceDictionaryItemsV1RemoteState(&service, name, dictName, expectedRemoteItems),
+					resource.TestCheckResourceAttr("fastly_service_dictionary_items_v1.items", "items.%", "2"),
+				),
+			},
+		},
+	})
+}
+
+// TestAccFastlyServiceDictionaryItemV1_create_inactive_service validates that
+// when creating a new inactive service consisting of a dictionary along with a
+// predefined list of items to populate it with, are applied successfully
+// instead of generating an error suggesting the dictionary ID was missing.
+//
+// NOTE: This error stemmed from a bug in our code (#345) where we discovered
+// that if a configuration has the activate field set to false AND it has no
+// previous active version, then the state wasn't being read. This manifested
+// itself as a runtime error in certain situations, such as another resource
+// referencing the state in its configuration.
+func TestAccFastlyServiceDictionaryItemV1_create_inactive_service(t *testing.T) {
+	var service gofastly.ServiceDetail
+	name := fmt.Sprintf("tf-test-%s", acctest.RandString(10))
+	dictName := fmt.Sprintf("dict %s", acctest.RandString(10))
+
+	expectedRemoteItems := map[string]string{
+		"key1": "value1",
+		"key2": "value2",
+	}
+
+	resource.Test(t, resource.TestCase{
+		PreCheck:     func() { testAccPreCheck(t) },
+		Providers:    testAccProviders,
+		CheckDestroy: testAccCheckServiceV1Destroy,
+		Steps: []resource.TestStep{
+			{
+				Config: testAccServiceDictionaryItemsV1Config_one_dictionary_with_items(name, dictName, expectedRemoteItems, false),
 				Check: resource.ComposeTestCheckFunc(
 					testAccCheckServiceV1Exists("fastly_service_v1.foo", &service),
 					testAccCheckFastlyServiceDictionaryItemsV1RemoteState(&service, name, dictName, expectedRemoteItems),
@@ -125,7 +162,7 @@ func TestAccFastlyServiceDictionaryItemV1_update(t *testing.T) {
 		CheckDestroy: testAccCheckServiceV1Destroy,
 		Steps: []resource.TestStep{
 			{
-				Config: testAccServiceDictionaryItemsV1Config_one_dictionary_with_items(name, dictName, expectedRemoteItems),
+				Config: testAccServiceDictionaryItemsV1Config_one_dictionary_with_items(name, dictName, expectedRemoteItems, true),
 				Check: resource.ComposeTestCheckFunc(
 					testAccCheckServiceV1Exists("fastly_service_v1.foo", &service),
 					testAccCheckFastlyServiceDictionaryItemsV1RemoteState(&service, name, dictName, expectedRemoteItems),
@@ -133,7 +170,7 @@ func TestAccFastlyServiceDictionaryItemV1_update(t *testing.T) {
 				),
 			},
 			{
-				Config: testAccServiceDictionaryItemsV1Config_one_dictionary_with_items(name, dictName, expectedRemoteItemsAfterUpdate),
+				Config: testAccServiceDictionaryItemsV1Config_one_dictionary_with_items(name, dictName, expectedRemoteItemsAfterUpdate, true),
 				Check: resource.ComposeTestCheckFunc(
 					testAccCheckServiceV1Exists("fastly_service_v1.foo", &service),
 					testAccCheckFastlyServiceDictionaryItemsV1RemoteState(&service, name, dictName, expectedRemoteItemsAfterUpdate),
@@ -156,7 +193,7 @@ func TestAccFastlyServiceDictionaryItemV1_external_item_is_removed(t *testing.T)
 		"key2": "value2",
 	}
 
-	config := testAccServiceDictionaryItemsV1Config_one_dictionary_with_items(name, dictName, expectedRemoteItems)
+	config := testAccServiceDictionaryItemsV1Config_one_dictionary_with_items(name, dictName, expectedRemoteItems, true)
 
 	resource.Test(t, resource.TestCase{
 		PreCheck:     func() { testAccPreCheck(t) },
@@ -204,7 +241,7 @@ func TestAccFastlyServiceDictionaryItemV1_external_item_deleted(t *testing.T) {
 		CheckDestroy: testAccCheckServiceV1Destroy,
 		Steps: []resource.TestStep{
 			{
-				Config: testAccServiceDictionaryItemsV1Config_one_dictionary_with_items(name, dictName, expectedRemoteItems),
+				Config: testAccServiceDictionaryItemsV1Config_one_dictionary_with_items(name, dictName, expectedRemoteItems, true),
 				Check: resource.ComposeTestCheckFunc(
 					testAccCheckServiceV1Exists("fastly_service_v1.foo", &service),
 					testAccCheckFastlyServiceDictionaryItemsV1RemoteState(&service, name, dictName, expectedRemoteItems),
@@ -244,7 +281,7 @@ func TestAccFastlyServiceDictionaryItemV1_batch_1001_items(t *testing.T) {
 		CheckDestroy: testAccCheckServiceV1Destroy,
 		Steps: []resource.TestStep{
 			{
-				Config: testAccServiceDictionaryItemsV1Config_one_dictionary_with_items(name, dictName, expectedRemoteItems),
+				Config: testAccServiceDictionaryItemsV1Config_one_dictionary_with_items(name, dictName, expectedRemoteItems, true),
 				Check: resource.ComposeTestCheckFunc(
 					testAccCheckServiceV1Exists("fastly_service_v1.foo", &service),
 					testAccCheckFastlyServiceDictionaryItemsV1RemoteState(&service, name, dictName, expectedRemoteItems),
@@ -273,7 +310,7 @@ func TestAccFastlyServiceDictionaryItemV1_import(t *testing.T) {
 		CheckDestroy: testAccCheckServiceV1Destroy,
 		Steps: []resource.TestStep{
 			{
-				Config: testAccServiceDictionaryItemsV1Config_one_dictionary_with_items(name, dictName, expectedRemoteItems),
+				Config: testAccServiceDictionaryItemsV1Config_one_dictionary_with_items(name, dictName, expectedRemoteItems, true),
 				Check: resource.ComposeTestCheckFunc(
 					testAccCheckServiceV1Exists("fastly_service_v1.foo", &service),
 				),
@@ -300,16 +337,15 @@ func testAccCheckFastlyServiceDictionaryItemsV1DoesNotExists(n string) resource.
 
 func testAccCheckFastlyServiceDictionaryItemsV1RemoteState(service *gofastly.ServiceDetail, name, dictName string, expectedItems map[string]string) resource.TestCheckFunc {
 	return func(s *terraform.State) error {
-
 		if service.Name != name {
 			return fmt.Errorf("Bad name, expected (%s), got (%s)", name, service.Name)
 		}
 
 		conn := testAccProvider.Meta().(*FastlyClient).conn
 		dict, err := conn.GetDictionary(&gofastly.GetDictionaryInput{
-			Service: service.ID,
-			Version: service.ActiveVersion.Number,
-			Name:    dictName,
+			ServiceID:      service.ID,
+			ServiceVersion: service.Version.Number,
+			Name:           dictName,
 		})
 
 		if err != nil {
@@ -317,8 +353,8 @@ func testAccCheckFastlyServiceDictionaryItemsV1RemoteState(service *gofastly.Ser
 		}
 
 		dictItems, err := conn.ListDictionaryItems(&gofastly.ListDictionaryItemsInput{
-			Service:    service.ID,
-			Dictionary: dict.ID,
+			ServiceID:    service.ID,
+			DictionaryID: dict.ID,
 		})
 
 		if err != nil {
@@ -346,8 +382,8 @@ func createDictionaryItemThroughApi(t *testing.T, service *gofastly.ServiceDetai
 	}
 
 	_, err = conn.CreateDictionaryItem(&gofastly.CreateDictionaryItemInput{
-		Service:    service.ID,
-		Dictionary: dict.ID,
+		ServiceID:    service.ID,
+		DictionaryID: dict.ID,
 
 		ItemKey:   expectedKey,
 		ItemValue: expectedValue,
@@ -363,14 +399,14 @@ func getDictionaryByName(service *gofastly.ServiceDetail, dictName string) (*gof
 	conn := testAccProvider.Meta().(*FastlyClient).conn
 
 	dict, err := conn.GetDictionary(&gofastly.GetDictionaryInput{
-		Service: service.ID,
-		Version: service.ActiveVersion.Number,
-		Name:    dictName,
+		ServiceID:      service.ID,
+		ServiceVersion: service.ActiveVersion.Number,
+		Name:           dictName,
 	})
 	return dict, err
 }
 
-func testAccServiceDictionaryItemsV1Config_one_dictionary_with_items(serviceName, dictName string, dictItemsList map[string]string) string {
+func testAccServiceDictionaryItemsV1Config_one_dictionary_with_items(serviceName, dictName string, dictItemsList map[string]string, activate bool) string {
 	backendName := fmt.Sprintf("%s.aws.amazon.com", acctest.RandString(3))
 	domainName := fmt.Sprintf("fastly-test.tf-%s.com", acctest.RandString(10))
 
@@ -386,7 +422,7 @@ func testAccServiceDictionaryItemsV1Config_one_dictionary_with_items(serviceName
 variable "mydict" {
 	type = object({ name=string, items=map(string) })
 	default = {
-		name = "%s" 
+		name = "%s"
 		items = %s
 	}
 }
@@ -408,6 +444,8 @@ resource "fastly_service_v1" "foo" {
 	name       = var.mydict.name
   }
 
+  activate = %t
+
   force_destroy = true
 }
 
@@ -415,7 +453,7 @@ resource "fastly_service_dictionary_items_v1" "items" {
     service_id = fastly_service_v1.foo.id
     dictionary_id = {for s in fastly_service_v1.foo.dictionary : s.name => s.dictionary_id}[var.mydict.name]
     items = var.mydict.items
-}`, dictName, dictItems, serviceName, domainName, backendName)
+}`, dictName, dictItems, serviceName, domainName, backendName, activate)
 }
 
 func testAccServiceDictionaryItemsV1Config_one_dictionary_no_items(serviceName, dictName string) string {